Convert knockout post to MDX
