--- content/posts/knockout-inline-confirm/.eslintrc.js
@@ -1,5 +1,0 @@
-module.exports = {
-  rules:{
-    "filenames/match-regex": "off"
-  }
-}
